% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.boi.finrep.data}
\alias{import.boi.finrep.data}
\title{Import BoI format financial report data}
\usage{
import.boi.finrep.data(filepath = NULL)
}
\arguments{
\item{filepath}{the path to financial report data (in csv format)}
}
\description{
This function imports financial report data from BOI format
}
