% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Import_functions.R
\name{import.TASE.comps_status}
\alias{import.TASE.comps_status}
\title{Import trading status data}
\usage{
import.TASE.comps_status(filepath = NULL)
}
\arguments{
\item{filepath}{the path to financial report data (in xlsx format)}
}
\description{
This function imports data set about listing status of TASE companies
}
